service, controller, request and response tested and working

+++ backend/src/main/java/com/ufes/miniproject/ProductRequestDTO.java
@@ -1,5 +1,5 @@
 package com.ufes.miniproject;
 
-public record ProductRequestDTO(String name, Float price, Integer amountInStore) {
+public record ProductRequestDTO(String name, Double price, Integer amountInStore) {
 
 }
